+++ lib/src/data/model/response/weather/weather_wind_model.dart
@@ -1,6 +1,6 @@
 class WeatherModelWind {
   final double? speed;
-  final double? deg;
+  final int? deg;
 
   WeatherModelWind({this.speed, this.deg});
